+++ core-ng/src/main/java/core/framework/internal/log/ActionLogMessageFactory.java
@@ -5,7 +5,6 @@ import core.framework.log.message.PerformanceStatMessage;
 import core.framework.util.Maps;
 import core.framework.util.Network;
 
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -13,7 +12,7 @@ import java.util.Map;
  */
 public class ActionLogMessageFactory {
     // 900K, kafka max message size is 1024 * 1024, leave 100K for rest, assume majority chars is in ascii (one char = one byte)
-    // refer to org.apache.kafka.clients.producer.ProducerConfig.MAX_REQUEST_SIZE_CONFIG
+    // refer to core.framework.internal.log.appender.KafkaAppender.append()
     private static final int HARD_TRACE_LIMIT = 900_000;
     // 600K assume majority chars is in ascii (one char = one byte)
     private static final int SOFT_TRACE_LIMIT = 600_000;
@@ -36,23 +35,9 @@ public class ActionLogMessageFactory {
         message.stats = log.stats;
         message.performanceStats = performanceStats(log.performanceStats);
         if (log.flushTraceLog()) {
-            // balance precision and cost, only estimate rough size limit
-            int contextSize = Math.min(estimatedSize(log.context), 300_000);    // leave some room for trace if context is too large
-            message.traceLog = log.trace(SOFT_TRACE_LIMIT - contextSize, HARD_TRACE_LIMIT - contextSize);
+            message.traceLog = log.trace(SOFT_TRACE_LIMIT, HARD_TRACE_LIMIT);
         }
         return message;
-    }
-
-    int estimatedSize(Map<String, List<String>> context) {
-        int size = 0;
-        for (Map.Entry<String, List<String>> entry : context.entrySet()) {
-            size += entry.getKey().length();
-            for (String value : entry.getValue()) {
-                // MessageListener may put null in context.key
-                if (value != null) size += value.length();
-            }
-        }
-        return size;
     }
 
     private Map<String, PerformanceStatMessage> performanceStats(Map<String, PerformanceStat> stats) {